refactor: import commands and evebts from __init__.py
