+++ pydoll/commands/__init__.py
@@ -1,4 +1,5 @@
 # global imports
+from pydoll.commands.browser import BrowserCommands
 from pydoll.commands.dom import DomCommands
 from pydoll.commands.fetch import FetchCommands
 from pydoll.commands.input import InputCommands
@@ -6,6 +7,7 @@ from pydoll.commands.network import NetworkCommands
 from pydoll.commands.page import PageCommands
 from pydoll.commands.runtime import RuntimeCommands
 from pydoll.commands.storage import StorageCommands
+from pydoll.commands.target import TargetCommands
 
 __all__ = [
     'DomCommands',
@@ -15,4 +17,6 @@ __all__ = [
     'PageCommands',
     'RuntimeCommands',
     'StorageCommands',
+    'BrowserCommands',
+    'TargetCommands',
 ]
