+++ tests/test_browser_window.py
@@ -2,7 +2,7 @@ import pytest
 import pytest_asyncio
 from unittest.mock import AsyncMock, MagicMock, patch
 from pydoll.browser.base import Browser
-from pydoll.commands.browser import BrowserCommands
+from pydoll.commands import BrowserCommands
 
 
 class ConcreteBrowser(Browser):
